Updated .bash_aliases and added a git_notes file
.bash_aliases was updated to include some of the commands used on
ng-hippocampus to enable various software packages for chip simulation
stuff. There are also some other updates to the alias file, like making
all grep commands colorized bu default, and some of the ls commands that
used to be in .bashrc are now in .bash_aliases

git_notes is a text file that has the shortcut commands for how to clone
the various git-repos in the ~/work/git-repos/.  It will save me future
time by not requiring me to make an internet search to find the proper
repository to clone from.