Modified .bashrc to not start tmux by default
This change will not start tmux by default until the bug with
vim is fixed.  This bug is where under tmux, vim no longer treats
meta keys the same as it used to under xterm.  (In tmux, the $TERM
variable is set to screen-256colors)  I should try some of the
stuff suggested at the following link:
  sunaku.github.io/vim-256color-bce.html