Added config files for Jupyter and IPython 4.x
Included in this are some of the customizations that Ben created to
enable notebook wide show/hide of input cells.

I am also now using code-folding provided by a git repository on the
internet, and as such, have included new notes in git-notes regarding
how to use that git repo.